+++ cli/internal/provider/testutils/example/model/writer.go
@@ -31,8 +31,9 @@ type RemoteWriter struct {
 // Metadata implements the RemoteResource interface
 func (r RemoteWriter) Metadata() handler.RemoteResourceMetadata {
 	return handler.RemoteResourceMetadata{
-		ID:         r.ID,
-		ExternalID: r.ExternalID,
-		Name:       r.Name,
+		ID:          r.ID,
+		ExternalID:  r.ExternalID,
+		WorkspaceID: backend.WorkspaceID,
+		Name:        r.Name,
 	}
 }